refactor(crawler): extract Engine struct with run() and shutdown()
